LPS-129801 Port remaining hook tests to TS
Tightened up the types for `useTimeout` while I was at it, because I
noticed that TS was inferring `any` in a couple of places.
