+++ modules/apps/frontend-js/frontend-js-react-web/test/js/hooks/useTimeout.tsx
@@ -15,45 +15,58 @@
 import {cleanup, fireEvent, render} from '@testing-library/react';
 import React from 'react';
 
-import useInterval from '../../../src/main/resources/META-INF/resources/js/hooks/useInterval';
+import useTimeout from '../../../src/main/resources/META-INF/resources/js/hooks/useTimeout';
 
-const {useEffect, useState} = React;
+const {useState} = React;
 
 const INTERVAL = 100;
 
-const Component = ({callback, onRender, onSchedule}) => {
-	const schedule = useInterval();
+type Callback = ReturnType<typeof jest.fn>;
+type Delay = ReturnType<typeof useTimeout>;
 
-	useEffect(() => {
-		if (onRender) {
-			onRender(schedule);
-		}
-	});
+const Component = ({
+	callback,
+	onDelay,
+	onRender,
+}: {
+	callback?: Callback;
+	onDelay?: (callback: () => void) => void;
+	onRender?: (delay: Delay) => void;
+}) => {
+	const delay = useTimeout();
 
-	const [, forceUpdate] = useState();
+	if (onRender) {
+		onRender(delay);
+	}
+
+	const [, forceUpdate] = useState(0);
 
 	const invoke = () => {
-		const cancel = schedule(callback, INTERVAL);
+		if (callback) {
+			const cancel = delay(callback, INTERVAL);
 
-		if (onSchedule) {
-			onSchedule(cancel);
+			if (onDelay) {
+				onDelay(cancel);
+			}
 		}
 	};
 
 	return (
 		<>
 			<button onClick={invoke}>Invoke</button>
-			<button onClick={forceUpdate}>Update</button>
+			<button onClick={() => forceUpdate((count) => count + 1)}>
+				Update
+			</button>
 		</>
 	);
 };
 
-describe('useInterval()', () => {
-	beforeEach(jest.useFakeTimers);
+describe('useTimeout()', () => {
+	beforeEach(() => jest.useFakeTimers());
 
 	afterEach(cleanup);
 
-	it('runs a function on a schedule', () => {
+	it('runs a function after a delay', () => {
 		const fn = jest.fn();
 
 		const {getByText} = render(<Component callback={fn} />);
@@ -62,41 +75,33 @@ describe('useInterval()', () => {
 
 		expect(fn).not.toHaveBeenCalled();
 
-		jest.advanceTimersByTime(INTERVAL);
+		jest.runAllTimers();
 
-		expect(fn).toHaveBeenCalledTimes(1);
-
-		jest.advanceTimersByTime(INTERVAL);
-
-		expect(fn).toHaveBeenCalledTimes(2);
+		expect(fn).toHaveBeenCalled();
 	});
 
-	it('returns a function that cancels the schedule', () => {
+	it('returns a function that cancels the timeout', () => {
 		const fn = jest.fn();
 
-		let cancel;
+		let cancel: () => void;
 
-		const onSchedule = (handle) => {
-			cancel = handle;
+		const onDelay = (callback: () => void): void => {
+			cancel = callback;
 		};
 
 		const {getByText} = render(
-			<Component callback={fn} onSchedule={onSchedule} />
+			<Component callback={fn} onDelay={onDelay} />
 		);
 
 		fireEvent.click(getByText('Invoke'));
 
 		expect(fn).not.toHaveBeenCalled();
 
-		jest.advanceTimersByTime(INTERVAL);
+		cancel!();
 
-		expect(fn).toHaveBeenCalledTimes(1);
+		jest.runAllTimers();
 
-		cancel();
-
-		jest.advanceTimersByTime(INTERVAL);
-
-		expect(fn).toHaveBeenCalledTimes(1);
+		expect(fn).not.toHaveBeenCalled();
 	});
 
 	it('does not run if unmounted', () => {
@@ -108,7 +113,7 @@ describe('useInterval()', () => {
 
 		unmount();
 
-		jest.advanceTimersByTime(INTERVAL);
+		jest.runAllTimers();
 
 		expect(fn).not.toHaveBeenCalled();
 	});
@@ -124,16 +129,16 @@ describe('useInterval()', () => {
 
 		render(<Component callback={fn} />);
 
-		jest.advanceTimersByTime(INTERVAL);
+		jest.runAllTimers();
 
 		expect(fn).not.toHaveBeenCalled();
 	});
 
-	it('preserves the identity of the schedule function', () => {
-		const functions = [];
+	it('preserves the identity of the delay function', () => {
+		const functions: Array<Delay> = [];
 
 		const {getByText} = render(
-			<Component onRender={(schedule) => functions.push(schedule)} />
+			<Component onRender={(delay) => functions.push(delay)} />
 		);
 
 		expect(functions.length).toBe(1);
